feat(security): add permission checks

+++ src/services/language.service.ts
@@ -1,23 +1,62 @@
 import { SessionUser } from "@variamosple/variamos-security";
+import { ensureCanCreate, ensureCanDelete, ensureCanRead, ensureCanUpdate } from "../auth/language.auth";
+import { Types } from "../models/language.model";
 import * as languageRepository from "../repositories/language.repository";
 import { LanguageQuery } from "../utils/language.query";
 
 export async function getLanguages(query: LanguageQuery, user: SessionUser) {
-    return languageRepository.findAll(query);
+    const languages = await languageRepository.findAll(query);
+    languages.forEach(language => {
+        ensureCanRead(user, language);
+    });
+    return languages;
 }
 
 export async function getLanguage(uuid: string, user: SessionUser) {
-    return languageRepository.findById(uuid);
+    const language = await languageRepository.findById(uuid);
+    if (!language) {
+        throw new Error("Language not found");
+    }
+    ensureCanRead(user, language);
+    return language;
 }
 
-export async function createLanguage(data: any, user: SessionUser) {
-    return languageRepository.create(data);
+export async function createLanguage(
+    data: {
+        name: string;
+        type: Types;
+    },
+    user: SessionUser,
+) {
+    ensureCanCreate(user);
+    return languageRepository.create({
+        ...data,
+        ownerId: user.id,
+    });
 }
 
-export async function updateLanguage(uuid: string, data: any, user: SessionUser) {
+export async function updateLanguage(
+    uuid: string,
+    data: Partial<{
+        name: string;
+        type: Types;
+        publicVersionId?: string;
+    }>,
+    user: SessionUser
+) {
+    const language = await languageRepository.findById(uuid);
+    if (!language) {
+        throw new Error("Language not found");
+    }
+    ensureCanUpdate(user, language);
     return languageRepository.update(uuid, data);
 }
 
 export async function deleteLanguage(uuid: string, user: SessionUser) {
+    const language = await languageRepository.findById(uuid);
+    if (!language) {
+        throw new Error("Language not found");
+    }
+    ensureCanDelete(user, language);
     return languageRepository.remove(uuid);
 }
